pyproject.toml
[project]
name = "tfm-2025"
version = "0.1.0"
description = "TFM para Master IA"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "langchain>=0.3.0",
    "langchain-ollama>=0.2.0",
    "pandas>=2.2.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
    "ruff>=0.8.0",
    "numpy>=1.26.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N"]
ignore = []
